requirements.txt
hydra-core==1.3.2
matplotlib==3.7.3
numpy==1.24.4
pandas == 1.5.3
omegaconf==2.3.0
pytorch_lightning==2.1.2
scikit_learn==1.3.1
timm==0.4.12
tqdm==4.66.1
torch-sparse==0.6.17
torch-scatter==2.1.1
pyg-lib @ https://data.pyg.org/whl/torch-2.1.0+cu118.html
torch-geometric==2.5.3
ogb==1.3.6
